backend/requirements.txt
alembic==1.16.4
fastapi==0.116.1
langchain==1.2.1
langchain-openai==1.1.12
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
PyMuPDF==1.26.3
python-docx==1.2.0
python-multipart==0.0.20
redis==6.4.0
SQLAlchemy==2.0.43
tenacity==9.1.2
uvicorn[standard]==0.35.0
